Updates the visual card when introducing another product
In this case, the Echo Show image is now displayed.
Q - "Does echo dot have a monitor?"
A - "No, but here's something I found on the Echo Show. Ask Alexa on the Echo Show to show you the front door or monitor the baby's room with compatible cameras from Amazon and others. Turn on lights or the TV, set thermostats, control Amazon Video on Fire TV, and more with WeMo, Philips Hue, Sony, ecobee, and other compatible smart home devices."